Swap snip and snap mappings for convenience
Since I use snippets more than simple placeholders, I have opted to put
M-h and M-l on previous and next node for luasnip and M-k for selecting
choice nodes, since it is between 'h' and 'l'.

Also, snapping uses M-i and M-o to jump to the previous and next snap
points, and M-p / M-P for placeholder / reminder is also a very
convenient mnemonic.